+++ lib/bmg/sql/relation.rb
@@ -167,7 +167,7 @@ module Bmg
       def _transform(type, transformation, options)
         expr = before_use(self.expr)
         sup, unsup = Processor::Transform.split_supported(transformation){|x|
-          [String, Integer, Float, Date, DateTime].include?(x)
+          [String, Integer, Float, Date, DateTime].include?(x) || x.is_a?(Hash)
         }
         return super if sup.nil?
         expr = Processor::Transform.new(sup, options, builder).call(expr)